Create useAddFriend custom hook
